Rename module to search_engine
